dev pull

+++ src/core/ScrollState.h
@@ -7,7 +7,7 @@ namespace scroll
 
 struct State
 {
-  std::uint8_t speedSetting = 4;      // 1–100
+  std::uint16_t speedSetting = 4;      // 1-500
   std::uint16_t textIntervalMs = 15;  // milliseconds per pixel step
   std::uint32_t lastScrollTickMs = 0; // Timestamp of last scroll update
   std::uint32_t lastBackgroundTickMs = 0; // Timestamp of last background update
